backend/requirements.txt
Django==5.0.3
djangorestframework==3.15.2
django-cors-headers==4.3.1
pandas==2.2.3
psycopg2-binary==2.9.9
python-dotenv==1.0.1
gunicorn==21.2.0
